backend/requirements.txt
pypdf
mistralai

python-dotenv
requests
pydantic
psycopg2-binary
fastapi
uvicorn
python-multipart